+++ db/migrate/20260404232000_create_message_contents.rb
@@ -15,57 +15,63 @@ class CreateMessageContents < ActiveRecord::Migration[8.1]
       FROM messages
     SQL
 
-    # Drop old FTS triggers (they reference messages.content_text)
-    execute "DROP TRIGGER IF EXISTS messages_au"
-    execute "DROP TRIGGER IF EXISTS messages_ad"
-    execute "DROP TRIGGER IF EXISTS messages_ai"
+    if connection.adapter_name == "SQLite"
+      # Drop old FTS triggers (they reference messages.content_text)
+      execute "DROP TRIGGER IF EXISTS messages_au"
+      execute "DROP TRIGGER IF EXISTS messages_ad"
+      execute "DROP TRIGGER IF EXISTS messages_ai"
+    end
 
     # Remove columns from messages
     remove_column :messages, :content_text
     remove_column :messages, :content_json
 
-    # Recreate FTS table pointing to message_contents as content source
-    execute "DROP TABLE IF EXISTS messages_fts"
-    execute <<~SQL
-      CREATE VIRTUAL TABLE messages_fts USING fts5(
-        content_text,
-        content='message_contents',
-        content_rowid='message_id',
-        tokenize='porter unicode61'
-      );
-    SQL
+    if connection.adapter_name == "SQLite"
+      # Recreate FTS table pointing to message_contents as content source
+      execute "DROP TABLE IF EXISTS messages_fts"
+      execute <<~SQL
+        CREATE VIRTUAL TABLE messages_fts USING fts5(
+          content_text,
+          content='message_contents',
+          content_rowid='message_id',
+          tokenize='porter unicode61'
+        );
+      SQL
 
-    # Populate FTS index from message_contents
-    execute <<~SQL
-      INSERT INTO messages_fts(rowid, content_text)
-      SELECT message_id, content_text FROM message_contents;
-    SQL
-
-    # Create triggers on message_contents
-    execute <<~SQL
-      CREATE TRIGGER message_contents_ai AFTER INSERT ON message_contents BEGIN
+      # Populate FTS index from message_contents
+      execute <<~SQL
         INSERT INTO messages_fts(rowid, content_text)
-        VALUES (new.message_id, new.content_text);
-      END;
+        SELECT message_id, content_text FROM message_contents;
+      SQL
 
-      CREATE TRIGGER message_contents_ad AFTER DELETE ON message_contents BEGIN
-        INSERT INTO messages_fts(messages_fts, rowid, content_text)
-        VALUES ('delete', old.message_id, old.content_text);
-      END;
+      # Create triggers on message_contents
+      execute <<~SQL
+        CREATE TRIGGER message_contents_ai AFTER INSERT ON message_contents BEGIN
+          INSERT INTO messages_fts(rowid, content_text)
+          VALUES (new.message_id, new.content_text);
+        END;
 
-      CREATE TRIGGER message_contents_au AFTER UPDATE ON message_contents BEGIN
-        INSERT INTO messages_fts(messages_fts, rowid, content_text)
-        VALUES ('delete', old.message_id, old.content_text);
-        INSERT INTO messages_fts(rowid, content_text)
-        VALUES (new.message_id, new.content_text);
-      END;
-    SQL
+        CREATE TRIGGER message_contents_ad AFTER DELETE ON message_contents BEGIN
+          INSERT INTO messages_fts(messages_fts, rowid, content_text)
+          VALUES ('delete', old.message_id, old.content_text);
+        END;
+
+        CREATE TRIGGER message_contents_au AFTER UPDATE ON message_contents BEGIN
+          INSERT INTO messages_fts(messages_fts, rowid, content_text)
+          VALUES ('delete', old.message_id, old.content_text);
+          INSERT INTO messages_fts(rowid, content_text)
+          VALUES (new.message_id, new.content_text);
+        END;
+      SQL
+    end
   end
 
   def down
-    execute "DROP TRIGGER IF EXISTS message_contents_au"
-    execute "DROP TRIGGER IF EXISTS message_contents_ad"
-    execute "DROP TRIGGER IF EXISTS message_contents_ai"
+    if connection.adapter_name == "SQLite"
+      execute "DROP TRIGGER IF EXISTS message_contents_au"
+      execute "DROP TRIGGER IF EXISTS message_contents_ad"
+      execute "DROP TRIGGER IF EXISTS message_contents_ai"
+    end
 
     add_column :messages, :content_text, :text
     add_column :messages, :content_json, :text
@@ -76,38 +82,40 @@ class CreateMessageContents < ActiveRecord::Migration[8.1]
         content_json = (SELECT content_json FROM message_contents WHERE message_contents.message_id = messages.id)
     SQL
 
-    # Restore original FTS table and triggers
-    execute "DROP TABLE IF EXISTS messages_fts"
-    execute <<~SQL
-      CREATE VIRTUAL TABLE messages_fts USING fts5(
-        content_text,
-        content='messages',
-        content_rowid='id',
-        tokenize='porter unicode61'
-      );
+    if connection.adapter_name == "SQLite"
+      # Restore original FTS table and triggers
+      execute "DROP TABLE IF EXISTS messages_fts"
+      execute <<~SQL
+        CREATE VIRTUAL TABLE messages_fts USING fts5(
+          content_text,
+          content='messages',
+          content_rowid='id',
+          tokenize='porter unicode61'
+        );
 
-      INSERT INTO messages_fts(rowid, content_text)
-      SELECT id, content_text FROM messages;
-    SQL
-
-    execute <<~SQL
-      CREATE TRIGGER messages_ai AFTER INSERT ON messages BEGIN
         INSERT INTO messages_fts(rowid, content_text)
-        VALUES (new.id, new.content_text);
-      END;
+        SELECT id, content_text FROM messages;
+      SQL
 
-      CREATE TRIGGER messages_ad AFTER DELETE ON messages BEGIN
-        INSERT INTO messages_fts(messages_fts, rowid, content_text)
-        VALUES ('delete', old.id, old.content_text);
-      END;
+      execute <<~SQL
+        CREATE TRIGGER messages_ai AFTER INSERT ON messages BEGIN
+          INSERT INTO messages_fts(rowid, content_text)
+          VALUES (new.id, new.content_text);
+        END;
 
-      CREATE TRIGGER messages_au AFTER UPDATE ON messages BEGIN
-        INSERT INTO messages_fts(messages_fts, rowid, content_text)
-        VALUES ('delete', old.id, old.content_text);
-        INSERT INTO messages_fts(rowid, content_text)
-        VALUES (new.id, new.content_text);
-      END;
-    SQL
+        CREATE TRIGGER messages_ad AFTER DELETE ON messages BEGIN
+          INSERT INTO messages_fts(messages_fts, rowid, content_text)
+          VALUES ('delete', old.id, old.content_text);
+        END;
+
+        CREATE TRIGGER messages_au AFTER UPDATE ON messages BEGIN
+          INSERT INTO messages_fts(messages_fts, rowid, content_text)
+          VALUES ('delete', old.id, old.content_text);
+          INSERT INTO messages_fts(rowid, content_text)
+          VALUES (new.id, new.content_text);
+        END;
+      SQL
+    end
 
     drop_table :message_contents
   end
